pyproject.toml
[project]
name = "bio-explainer"
version = "0.1.0"
description = "Local BYOK bio science explainer web app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "Pillow>=11.0",
    "openai>=1.60",
    "python-dotenv>=1.0",
    "pytest>=8.0",
    "httpx>=0.27",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
